+++ changes.md
@@ -1,13 +1,21 @@
 # OmegaT (v5.7.3) customization
 
-## Update 132_cs0 (2025-09-29)
+## Update 133_csp ()
 
+* Plugin: New `importExportCredentials-2.0.jar` to streamline team project authentication
+* Script: New `check_rules_CSV.groovy` with more checks and CSV output.
+* Script: Updated `scripts/project_changed/checkOMTversion.groovy` to run only for cApStAn builds of OmegaT in version 5.7.4
+* Script: Updated `scripts/updateConfigBundle.groovy` to disable running on non-cApStAn builds of OmegaT
+* Script: New `proj_custom_tags.groovy` to use project-specfic custom tag definitions
+
+## Update 132_cs0 (2026-01-05)
+
+* Config: Updated `omegat.prefs` with custom tags for EQLS 2026.
 * Script: Updated `write_project2tsv.groovy` to fix bug with handling ste.getDuplicate() as string (class org.omegat.core.data.SourceTextEntry$DUPLICATE')
 * Script: Added `utils_plunet_statsCSV.groovy`: Create Plunet Match Statistics in CSV
-* Script: Updated `scripts/project_changed/checkOMTversion.groovy` to improve the message to user and run only for cApStAn builds of OmegaT
+* Script: Updated `scripts/project_changed/checkOMTversion.groovy` to improve the message to user
 * Script: Removed `scripts/project_changed/utils_rename_project.groovy` which was buggy and not used
 * Script: Updated `scripts/project_changed/writeTMX4batch.groovy` to include support for PO files
-* Scruot: Updated `scripts/updateConfigBundle.groovy` to disable running on non-cApStAn builds of OmegaT
 
 ## Update 131_cs0 (2025-01-16)